Ignore irrelevant modifiers in key events.
Modifiers that change the key's symbol shouldn't be considered when
comparing keys, because they can vary between different keyboard
layouts. For example, the semicolon is shifted on DE keyboards and
isn't on US keyboards, so we should just use ';' for the binding and
ignore the shift modifier in the key event handler.

NOTE: Keybindings like 'S-x' in vimprobablerc will stop working and
      should now be written as simply 'X'.